refactor: move units.py from utils to calculations module
Relocates `units.py` from `apps/webapp/utils/` to
`apps/webapp/calculations/` to better reflect its usage context.
Updates all import references in `base.py`, `forms.py`, and `views.py`
to point to the new module location.
