+++ apps/webapp/forms.py
@@ -267,7 +267,7 @@ class SolutionForm(forms.Form):
                     f"The {label.lower()} cannot be equal or lower than zero."
                 )
 
-        from .utils.units import Q_
+        from .calculations.units import Q_
 
         if v1 is not None and v2 is not None:
             v1_q = Q_(v1, cleaned_data.get("v1_unit"))
