--- apps/webapp/utils/calculations.py
@@ -1,5 +1,0 @@
-"""Deprecated calculation helpers.
-
-All functions previously defined here are now available in
-``webapp.calculations.base``. Import and use the calculator classes instead.
-"""
